requirements.txt
# Core Python dependencies for Ethiopian Medical Business Data Platform

# Data Collection & Processing
requests==2.31.0
telethon==1.32.1
python-telegram-bot==20.7
beautifulsoup4==4.12.2
lxml==4.9.3

# Database & ORM
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Data Transformation & Analytics
pandas==2.1.4
numpy==1.24.4
dbt-core==1.7.3
dbt-postgres==1.7.3

# Machine Learning & Computer Vision
ultralytics==8.0.196
opencv-python==4.8.1.78
Pillow==10.1.0
torch==2.1.1
torchvision==0.16.1

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Orchestration
dagster==1.5.8
dagster-postgres==1.5.8

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Data Validation & Quality
great-expectations==0.17.23

# Logging & Monitoring
loguru==0.7.2

# HTTP Client
httpx==0.25.2

# JSON Processing
orjson==3.9.10

# Date/Time Handling
python-dateutil==2.8.2

# Async Support
asyncio-mqtt==0.16.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1 